+++ src/feature/auth/signup/ui/modal/index.tsx
@@ -1,0 +1,1 @@
+export { default as TermsModal } from './TermsModal';
